AI Backend/requirements.txt
# FastAPI core
fastapi
uvicorn

# LangChain ecosystem
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
langchain-google-genai
langchain-unstructured
langchain-exa
langchain-tavily
huggingface_hub[hf_xet]



# Vector DB support
faiss-cpu  # Optional: add if using FAISS

# Embedding models
sentence-transformers  # Required by HuggingFaceEmbeddings

# Document loading & parsing
unstructured[all-docs]
PyMuPDF
newspaper3k
lxml-html-clean  # fixed name from `lxml_html_clean`
beautifulsoup4
docx2txt
json5

# Utilities
python-dotenv
httpx
pydantic
ddgs  # DuckDuckGo search wrapper
firecrawl-py
pymupdf
